Split HIL tests into two files do and set/get test
