+++ Makefile
@@ -4,7 +4,7 @@ VERSION=0x03
 FTPUSER ?=joshc
 
 TST_FILES_ALL := $(wildcard tests/Test*.py)
-TST_FILES_LOCAL := $(filter-out tests/TestHIL.py, $(TST_FILES_ALL))
+TST_FILES_LOCAL := $(filter-out tests/TestHIL*, $(TST_FILES_ALL))
 
 .PHONY: test
 test:
@@ -12,7 +12,7 @@ test:
 
 .PHONY: test_HIL
 test_HIL:
-	@source venv/bin/activate && pytest --capture=tee-sys -vv tests/TestHIL.py
+	@source venv/bin/activate && pytest --capture=tee-sys -vv tests/TestHIL*
 
 .PHONY: test_all
 test_all:
@@ -30,7 +30,6 @@ run_test_server:
 transfer_working_files:
 	rsync -zvaP Makefile setup.py tests drops $(FTPUSER)@psbuild-rhel7:~/DoDRobotCode/
 
-# sync files to PSBUILD excluding any files from .gitignore
 .PHONY: sync
 sync:
 	rsync -zvaP --exclude-from='.gitignore' ../DropletDispensingRobot/ psbuild-rhel7:~/DoDRobotCode/
